requirements.txt
flask
werkzeug

torch
torchvision
torchaudio

segmentation-models-pytorch

numpy
opencv-python
Pillow
pyyaml

tqdm

matplotlib
seaborn
scikit-learn

scikit-image
